+++ common/src/main/kotlin/com/cobblemon/mod/common/entity/pokemon/ai/goals/PokemonInBattleMovementGoal.kt
@@ -13,6 +13,7 @@ import com.cobblemon.mod.common.battles.BattleSide
 import com.cobblemon.mod.common.entity.PoseType
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity
 import dev.architectury.event.events.common.TickEvent.Player
+import net.minecraft.command.argument.EntityAnchorArgumentType
 import net.minecraft.entity.LivingEntity
 import net.minecraft.entity.ai.TargetPredicate
 import net.minecraft.entity.ai.goal.Goal
@@ -34,8 +35,9 @@ class PokemonInBattleMovementGoal(val entity: PokemonEntity, val range: Int) : G
         val closestPokemonEntity = getClosestPokemonEntity()
         if (closestPokemonEntity != null) {
             entity.navigation.stop()
+//            entity.lookAt(EntityAnchorArgumentType.EntityAnchor.EYES, closestPokemonEntity.eyePos)
             entity.lookAtEntity(closestPokemonEntity, 30.0f, 30.0f)
-            entity.lookControl.lookAt(closestPokemonEntity.x, closestPokemonEntity.y, closestPokemonEntity.z)
+//            entity.lookControl.lookAt(closestPokemonEntity.x, closestPokemonEntity.y, closestPokemonEntity.z)
         }
     }
 }